Removed the possibility of not enabling the control over the DOOCS sigusr1/update routines when linking against this library. The function initialise() has been removed and the timing system is now always automatically initialised.
This change was introduced since the late initialisation of the timing system was always introducing race conditions which led to suprious failures (dead locks etc.) of tests using this library. If the control over the DOOCS timing is not wanted, the executable should no longer be linked against this library.

Since this is (intrinsically) an incompatible change, the major version has been increased along with this change.